% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_me_grad.R
\name{color_me_grad}
\alias{color_me_grad}
\title{color_me_grad}
\usage{
color_me_grad(u_data, info_overlay, your_palette)
}
\arguments{
\item{u_data}{dataframe holding output of dimension reduction}

\item{info_overlay}{numerical data that color_me_grad will represent with color gradations}

\item{your_palette}{the palette of your choice (please see colourvalues documentation to see options)}
}
\value{
dataframe with original dimension reduction and column with color values, as well as the info_overlay column
}
\description{
function to make color info dataframe corresponding to your feature of interest for your data that you want represented via color, then adding it to the dimension reduced dataframe
}
\examples{
color_me_grad(umakea(iris[,1:4]), iris$Sepal.Length, "viridis")
}
\seealso{
\code{\link[colourvalues]{colour_values}}
}
