% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact_multi.R
\name{interact_multi}
\alias{interact_multi}
\title{interact_multi}
\usage{
interact_multi(count_data, main_info_data, sec_info_data, dim_red_meth,
  your_palette, main_info_name, sec_info_name)
}
\arguments{
\item{count_data}{PARAM_DESCRIPTION}

\item{main_info_data}{PARAM_DESCRIPTION}

\item{sec_info_data}{PARAM_DESCRIPTION}

\item{dim_red_meth}{PARAM_DESCRIPTION}

\item{your_palette}{PARAM_DESCRIPTION}

\item{main_info_name}{PARAM_DESCRIPTION}

\item{sec_info_name}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
