% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact_qual.R
\name{interact_qual}
\alias{interact_qual}
\title{interact_qual}
\usage{
interact_qual(count_data, qual_data, dim_red_meth = c("UMAP", "PCA"))
}
\arguments{
\item{count_data}{dataframe of input data}

\item{qual_data}{qualitative data used to color points of plot}

\item{dim_red_meth}{dimension reduction method: "pca" or "umap" are the options available now}
}
\value{
an interactive dimension reduction plot that is color-coordinated with qualitative data, and on hover, displays sample name and the qualitative feature of interest
}
\description{
Generates an interactive dimension reduction plot that is color-coordinated with qualitative data related to your dataset.
}
\examples{
\dontrun{
if(interactive()){
 interact_qual(iris[,1:4], iris$Species, "UMAP")
 }
}
}
