% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact_quant.R
\name{interact_quant}
\alias{interact_quant}
\title{interact_quant}
\usage{
interact_quant(count_data, quant_info_name, quant_info, dim_red_meth,
  your_palette)
}
\arguments{
\item{count_data}{PARAM_DESCRIPTION}

\item{quant_info_name}{PARAM_DESCRIPTION}

\item{quant_info}{PARAM_DESCRIPTION}

\item{dim_red_meth}{PARAM_DESCRIPTION}

\item{your_palette}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
