% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactmapper-package.R
\docType{package}
\name{interactmapper-package}
\alias{interactmapper}
\alias{interactmapper-package}
\title{interactmapper: An Interactive Plotting Tool for Dimension Reduction}
\description{
The package interactmapper makes interpreting dimension reduction results more intuitive and open to exploration, by providing easy ways to overlay secondary information on your plotted data, with hovertext labels featuring the features of your choice, color coordinated based on your primary feature of interest.
}
\author{
\strong{Maintainer}: Octavia Dancu-Lixandru \email{octavia-maria.dancu@mail.mcgill.ca} (YOUR-ORCID-ID)

}
\keyword{internal}
