% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_maker.R
\name{pca_maker}
\alias{pca_maker}
\title{pca_maker}
\usage{
pca_maker(count_data)
}
\arguments{
\item{count_data}{A dataframe holding the data that you want to apply umap to}
}
\value{
a dataframe with the two feature columns resulting from the PCA dimension reduction and a sample names column
}
\description{
This function takes in your count data and returns a dataframe with the two feature columns resulting from the PCA dimension reduction and a sample names column
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
