% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umakea.R
\name{umakea}
\alias{umakea}
\title{umakea}
\usage{
umakea(count_data)
}
\arguments{
\item{count_data}{A dataframe with the input data}
}
\value{
a dataframe with the two feature columns resulting from the UMAP dimension reduction and a sample names column
}
\description{
This function takes in your count data and returns a dataframe with the two feature columns resulting from the UMAP dimension reduction and a sample names column
}
\examples{
p<-umakea(iris[,1:4])
}
\seealso{
\code{\link[umap]{umap}}
}
