% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_me_grad.R
\name{color_me_grad}
\alias{color_me_grad}
\title{color_me_grad}
\usage{
color_me_grad(u_data, info_overlay, your_palette)
}
\arguments{
\item{u_data}{dataframe holding output of dimension reduction}

\item{info_overlay}{numerical data that color_me_grad will represent with color gradations}

\item{your_palette}{the palette of your choice (please see colourvalues documentation to see options)}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
function to make color info dataframe corresponding to DNA methylation, then adding it to the dimension reduced dataframe
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[colourvalues]{colour_values}}
}
