% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact_qual.R
\name{interact_qual}
\alias{interact_qual}
\title{interact_qual}
\usage{
interact_qual(count_data, qual_data, dim_red_meth)
}
\arguments{
\item{count_data}{PARAM_DESCRIPTION}

\item{qual_data}{PARAM_DESCRIPTION}

\item{dim_red_meth}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
