% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_maker.R
\name{pca_maker}
\alias{pca_maker}
\title{pca_maker}
\usage{
pca_maker(count_data)
}
\arguments{
\item{count_data}{A dataframe with the input data}
}
\value{
a dataframe with the two feature columns resulting from the PCA dimension reduction and a sample names column
}
\description{
Returns a dataframe with the two feature columns resulting from the PCA (principal components analysis) dimension reduction and a sample names column
}
\examples{
\dontrun{
if(interactive()){
 p<-pca_maker(iris[,1:4])
 }
}

}
